requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-timezone-field==4.0
djangorestframework==3.11.0
idna==2.8
kombu==4.6.7
psycopg2==2.8.4
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2019.3
redis==3.3.11
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
vine==1.3.0